requirements.txt
Django==2.2.4
django-cors-headers==3.0.2
django-stubs==1.0.2
djangorestframework==3.10.2
djangorestframework-stubs==1.0.0
mypy==0.720
mypy-extensions==0.4.1
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
typed-ast==1.4.0
typing-extensions==3.7.4
